airflow/requirements.txt
pandas
faker
google-cloud-bigquery
google-auth
